+++ src/python/arcor2_calibration/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## [0.8.0] - 2022-07-07
+
+### Changed
+
+- **BREAKING**: Implement new error handling flow. Error codes of **every** endpoint were replaced with error 
+  types as described in swagger documentation.
+
 ## [0.7.0] - 2021-10-25
 
 ### Changed