+++ src/python/arcor2_calibration_data/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## [0.5.0] - 2022-07-07
+### Added
+- Service exceptions added.
+
 ## [0.4.0] - 2021-06-11
 ### Changed
 - Client now raises custom exceptions.